renamed module for return_mode
