+++ src/lib.rs
@@ -28,7 +28,7 @@ mod salsa_struct;
 mod storage;
 mod table;
 mod tracked_struct;
-mod traits;
+mod return_mode;
 mod update;
 mod views;
 mod zalsa;
@@ -53,8 +53,8 @@ pub use self::key::DatabaseKeyIndex;
 pub use self::revision::Revision;
 pub use self::runtime::Runtime;
 pub use self::storage::{Storage, StorageHandle};
-pub use self::traits::SalsaAsRef;
-pub use self::traits::SalsaAsDeref;
+pub use self::return_mode::SalsaAsRef;
+pub use self::return_mode::SalsaAsDeref;
 pub use self::update::Update;
 pub use self::zalsa::IngredientIndex;
 pub use crate::attach::{attach, with_attached_database};
